+++ src/model/point.rs
@@ -1,5 +1,5 @@
 #[derive(Debug)]
 pub struct Point {
-    pub x: u32,
-    pub y: u32,
+    pub x: usize,
+    pub y: usize,
 }
